+++ src/wai/annotations/imgaug/isp/linear_contrast/component/_LinearContrast.py
@@ -21,6 +21,15 @@ class LinearContrast(BaseImageAugmentation):
         help="the maximum alpha to apply to the images"
     )
 
+    def _default_suffix(self):
+        """
+        Returns the default suffix to use for images when using "add" rather than "replace" as mode.
+
+        :return: the default suffix
+        :rtype: str
+        """
+        return "-contrasted"
+
     def _can_augment(self):
         """
         Checks whether augmentation can take place.