+++ src/wai/annotations/imgaug/isp/rotate/component/_Rotate.py
@@ -22,6 +22,15 @@ class Rotate(BaseImageAugmentation):
         help="the end of the degree range to use for rotating the images"
     )
 
+    def _default_suffix(self):
+        """
+        Returns the default suffix to use for images when using "add" rather than "replace" as mode.
+
+        :return: the default suffix
+        :rtype: str
+        """
+        return "-rotated"
+
     def _can_augment(self):
         """
         Checks whether augmentation can take place.