+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+1.0.1 (????-??-??)
+------------------
+
+- added more typing hints to parameters
+- added ability to add augmented images rather than replace original ones, using
+  the `-m/--mode add` option (`--suffix` overrides the default suffix)
+
+
 1.0.0 (2022-02-23)
 ------------------